fix: resolve compilation errors in gen UI components
Fixed two compilation errors:

1. gen_ui_chart.dart: Updated tooltipBgColor to getTooltipColor
   - fl_chart library uses getTooltipColor callback instead of tooltipBgColor
   - Changed from: tooltipBgColor: Color
   - Changed to: getTooltipColor: (BarChartGroupData group) => Color

2. gen_ui_progress.dart: Fixed percentage type casting
   - percentage needs to be explicitly cast to double
   - Changed clamp(0, 100) to clamp(0.0, 100.0).toDouble()
   - Ensures _getProgressColor receives double parameter

Build now succeeds without errors.

+++ lib/features/chat/presentation/widgets/gen_ui/gen_ui_progress.dart
@@ -11,7 +11,7 @@ class GenUIProgress extends StatelessWidget {
     final title = data['title'] as String;
     final current = (data['current'] as num).toDouble();
     final target = (data['target'] as num).toDouble();
-    final percentage = (current / target * 100).clamp(0, 100);
+    final percentage = (current / target * 100).clamp(0.0, 100.0).toDouble();
     final label = data['label'] as String?;
     final showValue = data['showValue'] as bool? ?? true;
